+++ GUI/MainMenu.h
@@ -12,7 +12,7 @@
 #include <filesystem>
 #include <fstream>
 #include <iostream>
-#include "../ProjectSourceCode/Map.h"
+#include "../ProjectSourceCode/Map/Map.h"
 #include "../ProjectSourceCode/Builder/MapBuilder.h"
