+++ [COMP345] Assignment 1.vcxproj
@@ -113,7 +113,7 @@
     <ClCompile Include="ProjectSourceCode\Test\testitem.cpp" />
     <ClCompile Include="ProjectSourceCode\Test\testitemcontainer.cpp" />
     <ClCompile Include="ProjectSourceCode\Dice\Dice.cpp" />
-    <ClCompile Include="ProjectSourceCode\Map.cpp" />
+    <ClCompile Include="ProjectSourceCode\Map\Map.cpp" />
     <ClCompile Include="ProjectSourceCode\Serialize\serializeItem.cpp" />
     <ClCompile Include="ProjectSourceCode\Test\testSerializeCharacter.cpp" />
     <ClCompile Include="ProjectSourceCode\Test\testserializeitem.cpp" />
@@ -157,7 +157,7 @@
     <ClInclude Include="ProjectSourceCode\Character\Character.h" />
     <ClInclude Include="ProjectSourceCode\Dice\Dice.h" />
     <ClInclude Include="ProjectSourceCode\Test\testitemcontainer.h" />
-    <ClInclude Include="ProjectSourceCode\Map.h" />
+    <ClInclude Include="ProjectSourceCode\Map\Map.h" />
     <ClInclude Include="ProjectSourceCode\Serialize\serializeItem.h" />
     <ClInclude Include="ProjectSourceCode\Test\testSerializeCharacter.h" />
     <ClInclude Include="ProjectSourceCode\Test\testserializeitem.h" />
